# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document Search API",
        "description": "A powerful document search API engine using TNT Search for fast full-text search capabilities. This API supports multi-tenant workspaces, document upload, text extraction, and advanced search features.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "List all documents",
                "description": "Get a paginated list of all documents in the workspace",
                "operationId": "getDocuments",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filter by project ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Document"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a new document",
                "description": "Upload a document file to the workspace and automatically extract text content for search indexing",
                "operationId": "storeDocument",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Document file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "description": "Optional document title",
                                        "type": "string",
                                        "example": "Quarterly Report"
                                    },
                                    "project_id": {
                                        "description": "Optional project ID to associate with document",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Document"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
...
        "/api/documents/{id}/download": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Download document file",
                "description": "Download the original file of a document",
                "operationId": "downloadDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document file download",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document or file not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/documents-statistics": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get document statistics",
                "description": "Get statistical information about documents in the workspace",
                "operationId": "getDocumentStatistics",
                "responses": {
                    "200": {
                        "description": "Document statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "total_documents": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "indexed_documents": {
                                                            "type": "integer",
                                                            "example": 145
                                                        },
                                                        "total_size": {
                                                            "type": "integer",
                                                            "example": 52428800
                                                        },
                                                        "formatted_total_size": {
                                                            "type": "string",
                                                            "example": "50 MB"
                                                        },
                                                        "file_types": {
                                                            "type": "object"
                                                        },
                                                        "recent_uploads": {
                                                            "type": "integer",
                                                            "example": 15
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/documents-bulk-index": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Bulk index documents",
                "description": "Index all unindexed documents in the workspace for search",
                "operationId": "bulkIndexDocuments",
                "responses": {
                    "200": {
                        "description": "Documents indexed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "indexed_count": {
                                                            "type": "integer",
                                                            "example": 25
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/index/status": {
            "get": {
                "tags": [
                    "Index"
                ],
                "summary": "Get search index status",
                "description": "Get the current status of the search index for the workspace",
                "operationId": "getIndexStatus",
                "responses": {
                    "200": {
                        "description": "Index status information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "indexed_documents": {
                                                            "type": "integer",
                                                            "example": 145
                                                        },
                                                        "total_documents": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "index_size": {
                                                            "type": "integer",
                                                            "example": 2048576
                                                        },
                                                        "formatted_index_size": {
                                                            "type": "string",
                                                            "example": "2 MB"
                                                        },
                                                        "last_updated": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "index_health": {
                                                            "type": "string",
                                                            "enum": [
                                                                "healthy",
                                                                "warning",
                                                                "error"
                                                            ]
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/index/rebuild": {
            "post": {
                "tags": [
                    "Index"
                ],
                "summary": "Rebuild search index",
                "description": "Rebuild the entire search index for the workspace",
                "operationId": "rebuildIndex",
                "responses": {
                    "200": {
                        "description": "Index rebuilt successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "indexed_documents": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "processing_time": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 45.23
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
...
        "/api/projects/{id}/statistics": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project statistics",
                "description": "Get detailed statistics for a specific project",
                "operationId": "getProjectStatistics",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "total_documents": {
                                                            "type": "integer"
                                                        },
                                                        "indexed_documents": {
                                                            "type": "integer"
                                                        },
                                                        "unindexed_documents": {
                                                            "type": "integer"
                                                        },
                                                        "total_size": {
                                                            "type": "integer"
                                                        },
                                                        "formatted_total_size": {
                                                            "type": "string"
                                                        },
                                                        "recent_uploads": {
                                                            "type": "integer"
                                                        },
                                                        "file_types": {
                                                            "type": "object"
                                                        },
                                                        "last_indexed_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Document": {
                "required": [
                    "id",
                    "title",
                    "file_type",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Annual Report 2024"
                    },
                    "file_type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 2048576
                    },
                    "formatted_file_size": {
                        "type": "string",
                        "example": "2 MB"
                    },
                    "original_filename": {
                        "type": "string",
                        "example": "annual_report_2024.pdf"
                    },
                    "indexed": {
                        "type": "boolean",
                        "example": true
                    },
                    "indexed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "WorkspaceKey": {
                "type": "apiKey",
                "description": "Workspace API key for authentication",
                "name": "X-Workspace-Key",
                "in": "header"
            }
        }
    },
...
}